+++ build.gradle
@@ -45,7 +45,7 @@ def dependencyVersions = [
         ida_test_utils:"2.0.0-46",
         opensaml:"$opensaml_version",
         dev_pki: '1.1.0-37',
-        saml_lib:"$opensaml_version-227"
+        saml_lib:"$opensaml_version-230"
 ]
 
 subprojects {
