+++ hub-saml/src/main/java/uk/gov/ida/saml/hub/transformers/inbound/providers/DecoratedSamlResponseToIdaResponseIssuedByIdpTransformer.java
@@ -1,8 +1,8 @@
 package uk.gov.ida.saml.hub.transformers.inbound.providers;
 
 import org.opensaml.saml.saml2.core.Response;
+import uk.gov.ida.eidas.logging.EidasAttributesLogger;
 import uk.gov.ida.saml.core.validators.DestinationValidator;
-import uk.gov.ida.saml.hub.domain.EidasAttributesLogger;
 import uk.gov.ida.saml.hub.domain.InboundResponseFromIdp;
 import uk.gov.ida.saml.hub.transformers.inbound.IdaResponseFromIdpUnmarshaller;
 import uk.gov.ida.saml.hub.validators.response.idp.IdpResponseValidator;
